✨ feat(shop): add shop feature
This commit introduces a new feature that allows adding products to a shop.

The following changes were made:

- Created a new Shop model and migration.
- Added a new ShopController with endpoints for adding and removing products from the shop.
- Added a new ShopResource for displaying shop data.
- Updated the Product model to include a relationship with the Shop model.
- Updated the ProductController to prevent deleting products that are in the shop.
- Added a new ShopFactory for creating test data.
- Added new API routes for the shop feature.

+++ app/Http/Controllers/Api/ProductController.php
@@ -130,21 +130,31 @@ class ProductController extends Controller
      * Delete Product
      *
      * @param  $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse|\Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $productInUsed = false;
 
         $product = Products::findOrFail($id);
-        // NOTE: Check and make sure that the product will not be deleted if it in use by other module
+
         if (Wishlist::has($product)) {
             $productInUsed = true;
         }
-        // TODO: Add checking if the product is in used in the Shop/Store
+
+        if ($product->shop()->count() > 0) {
+            $productInUsed = true;
+        }
+
+        if ($productInUsed) {
+            return response()->json([
+                'message' => 'Unable to delete product as it is still in use',
+                'success' => false,
+            ], Response::HTTP_FORBIDDEN);
+        }
 
         $product->delete();
 
-        return response()->json(status: Response::HTTP_NO_CONTENT);
+        return response()->noContent();
     }
 }
